Remove scaling of gravity-wave absorbing layer coefficient by local mesh size
This commit removes the scaling of the gravity-wave absorbing layer coefficient
by the local mesh size.  Previously, the absorbing layer coefficient was scaled
by dx/dx_fine.  For large values of dx/dx_fine instabilities have been
encountered associated with this scaling of the absorbing layer coefficient.

Units have been added to the configuration variable config_xnutr (1/s) in the
Registry - it is used in calculating the absorbing layer coefficient.  Units
have also been added to config_visc4_2dsmag (m/s) in the Registry - it is used
in the calculation of the hyperviscosity.